+++ inc/hooks.php
@@ -6,8 +6,8 @@
  *
  * @Author:		Roni Laukkarinen
  * @Date:   		2022-06-30 16:24:47
- * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2022-06-30 16:36:05
+ * @Last Modified by:   Timi Wahalahti
+ * @Last Modified time: 2022-09-22 12:14:52
  *
  * @package air-light
  */
@@ -49,7 +49,6 @@ add_filter( 'allowed_block_types_all', __NAMESPACE__ . '\allowed_block_types', 1
 add_filter( 'use_block_editor_for_post_type', __NAMESPACE__ . '\use_block_editor_for_post_type', 10, 2 );
 add_action( 'enqueue_block_editor_assets', __NAMESPACE__ . '\register_block_editor_assets' );
 add_filter( 'block_editor_settings_all', __NAMESPACE__ . '\remove_gutenberg_inline_styles', 10, 2 );
-add_action( 'after_setup_theme', __NAMESPACE__ . '\setup_editor_styles' );
 add_action( 'enqueue_block_editor_assets', __NAMESPACE__ . '\block_editor_title_input_styles' );
 
 /**