+++ js/src/front-end.js
@@ -8,6 +8,7 @@ import reframe from 'reframe.js';
 import { styleExternalLinks, initExternalLinkLabels } from './modules/external-link';
 import initAnchors from './modules/anchors';
 import backToTop from './modules/top';
+import initA11ySkipLink from './modules/a11y-skip-link';
 import 'what-input';
 import './modules/navigation';
 // import './modules/sticky-nav';
@@ -21,6 +22,7 @@ document.addEventListener('DOMContentLoaded', () => {
   backToTop();
   styleExternalLinks();
   initExternalLinkLabels();
+  initA11ySkipLink();
 
   // Fit video embeds to container
   reframe('.wp-has-aspect-ratio iframe');