+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### [Unreleased]
+### 9.1.7: 2022-09-28
 
 * Add external-link-icon styles for mobile
 * Remove setup_editor_styles function because it breaks stage+prod and block styles still load correctly without that